バグ対応&UIの調整
